further improvements on indicators
